Cache `_get_e2e_cross_signing_signatures_for_devices`
The use of `@cachedList` requires having an cached, equivalent "single
item" method that represents the key and cached value. Calls to methods
with @cachedList are then processed and items that are already cached
are removed before the @cachedList-decorated method is called. The
single item method should not be directly used.

The return type needed to be changed to a "Sequence" to satisfy the
requirement that cached results must be immutable (lest you risk
mutating the cache!).